Added fixes

+++ lesson1/S_treugolnika.rb
@@ -1,7 +1,7 @@
 puts 'Введите длину основания'
-a = gets.to_i
+a = gets.to_f
 
 puts 'Введите длину высоты'
-h = gets.to_i
+h = gets.to_f
 
 puts "Площадь #{a*h/2}"
